pyproject.toml
[tool.poetry]
name = "sism"
version = "0.1.0"
description = ""
authors = ["David Kellerman <david@silvermoon.org>"]

[tool.poetry.dependencies]
python = "3.9.13"
Django = "^4.0.4"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
django-on-heroku = "^1.1.2"
nltk = "^3.7"
eng-to-ipa = "^0.0.2"
pronouncing = "^0.2.0"
inflect = "^6.0.2"
num2words = "^0.5.10"
panphon = "^0.20.0"
g2p = "^1.1.20230511"
minineedle = "^3.1.5"
django-cors-headers = "^3.12.0"
django-debug-toolbar = "^3.4.0"
requests = "^2.28.0"
requests-cache = "^0.9.4"
hiredis = "^2.0.0"
redis = "^4.3.3"


[tool.poetry.group.dev.dependencies]
graphene-django = "^3.0.0"
django-graphql-jwt = "^0.3.4"
tqdm = "^4.64.0"
django-storages = {extras = ["google"], version = "^1.12.3"}
pafy = "^0.5.5"
gruut-ipa = "^0.13.0"
django-smuggler = "^1.0.4"
django-pandas = "^0.6.6"
youtube_dl = "2020.12.2"
ffmpeg = "^1.4"
pytest = "^7.2"
torch = "^2.0.1"
scikit-learn = "^1.1.1"
openai = "^0.27.8"
sh = "^1.14.2"
autopep8 = "^2.0.0"
gensim = "^4.2.0"
numpy = "^1.22.4"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
